src/requirements.txt
Django==1.6.2
Pillow==2.3.0
django-extensions==1.1.1
django-guardian==1.1.1
django-multiuploader==0.2.6
django-social-auth==0.7.25
django-stdimage==0.4.2
django-userena==1.3.0
easy-thumbnails==1.4
html2text==3.200.3
httplib2==0.8
oauth2==1.5.211
pycrypto==2.6.1
pyelasticsearch==0.5
pyjade==2.0.3
python-magic==0.4.6
python-openid==2.2.5
requests==1.2.3
simplejson==3.3.2
six==1.5.2
sorl-thumbnail==11.12
stripe==1.12.0
wsgiref==0.1.2